refactor(test): Make it easier to fork tests
