+++ tests/builder/app_settings.rs
@@ -1,6 +1,6 @@
 use std::ffi::OsString;
 
-use crate::utils;
+use super::utils;
 
 use clap::{arg, error::ErrorKind, AppSettings, Arg, Command};
