+++ tests/builder/arg_aliases_short.rs
@@ -1,4 +1,4 @@
-use crate::utils;
+use super::utils;
 
 use clap::{arg, Arg, Command};
