+++ tests/builder/default_vals.rs
@@ -1,4 +1,4 @@
-use crate::utils;
+use super::utils;
 use clap::{arg, error::ErrorKind, Arg, Command};
 
 #[test]
