+++ tests/builder/derive_order.rs
@@ -1,4 +1,4 @@
-use crate::utils;
+use super::utils;
 
 use std::str;
