+++ tests/builder/error.rs
@@ -1,4 +1,4 @@
-use crate::utils;
+use super::utils;
 
 use clap::{arg, error::ErrorKind, value_parser, Arg, Command, Error};
