+++ tests/builder/flags.rs
@@ -1,4 +1,4 @@
-use crate::utils;
+use super::utils;
 use clap::{arg, Arg, Command};
 
 const USE_FLAG_AS_ARGUMENT: &str =
