+++ tests/builder/help.rs
@@ -1,4 +1,4 @@
-use crate::utils;
+use super::utils;
 
 use clap::{arg, error::ErrorKind, Arg, ArgGroup, Command, PossibleValue};
