+++ tests/builder/hidden_args.rs
@@ -1,4 +1,4 @@
-use crate::utils;
+use super::utils;
 
 use clap::{arg, Arg, Command};
