+++ tests/builder/possible_values.rs
@@ -1,4 +1,4 @@
-use crate::utils;
+use super::utils;
 
 use clap::{error::ErrorKind, Arg, Command, PossibleValue};
