+++ tests/builder/tests.rs
@@ -1,4 +1,4 @@
-use crate::utils;
+use super::utils;
 
 use std::io::Write;
 use std::str;
