+++ tests/builder/version.rs
@@ -1,4 +1,4 @@
-use crate::utils;
+use super::utils;
 
 use std::str;
